removed foxglove and topicbus, then split files for smaller files
