+++ helios_sim/src/prelude.rs
@@ -20,7 +20,7 @@ pub use crate::simulation::core::components::{
 };
 pub use crate::simulation::core::components::{EstimationDynamicsModel, MeasurementModel};
 pub use crate::simulation::core::spawn_requests::SpawnAgentConfigRequest;
-pub use crate::simulation::core::transforms::{TfFramePose, TfTree, TrackedFrame};
+pub use crate::simulation::core::transforms::{TfTree, TrackedFrame};
 
 pub use crate::simulation::plugins::sensors::imu::ImuPlugin;
 pub use crate::simulation::plugins::vehicles::ackermann::{
