+++ helios_sim/src/simulation/core/simulation_setup.rs
@@ -12,12 +12,8 @@ use crate::simulation::core::app_state::{AssetLoadSet, SimulationSet};
 use crate::simulation::core::events::BevyMeasurementMessage;
 use crate::simulation::core::prng::SimulationRng;
 use crate::simulation::core::transforms::build_static_tf_maps;
-use crate::simulation::core::{
-    ground_truth_publish_system, ground_truth_sync_system, tf_publish_system,
-};
-// Import the Bevy-specific types this plugin manages
+use crate::simulation::core::ground_truth_sync_system;
 use super::components::GroundTruthState;
-use super::topics::TopicBus;
 use super::transforms::{tf_tree_incremental_update_system, tf_tree_structural_system, TfTree};
 
 pub struct SimulationSetupPlugin;
@@ -41,8 +37,7 @@ impl Plugin for SimulationSetupPlugin {
         app.insert_resource(SimulationRng(rng));
 
         // --- INITIALIZE RESOURCES & EVENTS ---
-        app.init_resource::<TopicBus>()
-            .init_resource::<TfTree>()
+        app.init_resource::<TfTree>()
             .add_event::<BevyMeasurementMessage>();
 
         let fixed_update_timestep = 1.0 / frequency_hz;
@@ -168,12 +163,6 @@ impl Plugin for SimulationSetupPlugin {
                 tf_tree_incremental_update_system
                     .in_set(SimulationSet::StateSync)
                     .after(ground_truth_sync_system)
-                    .run_if(in_state(AppState::Running)),
-                ground_truth_publish_system
-                    .in_set(SimulationSet::Validation)
-                    .run_if(in_state(AppState::Running)),
-                tf_publish_system
-                    .in_set(SimulationSet::Validation)
                     .run_if(in_state(AppState::Running)),
             ),
         );
