+++ helios_sim/src/simulation/mod.rs
@@ -1,6 +1,6 @@
 //! Top-level simulation module grouping all ECS subsystems.
 //!
-//! Submodules: `core` (TF tree, TopicBus, ground-truth sync), `plugins` (sensor, vehicle,
+//! Submodules: `core` (TF tree, ground-truth sync), `plugins` (sensor, vehicle,
 //! autonomy, control, debugging plugins), `profile` + `profile_plugin` (capability-gated
 //! plugin loading), `registry` (AutonomyRegistry factory dispatch), `config`, `utils`.
