+++ helios_sim/src/simulation/plugins/autonomy/systems/spawn.rs
@@ -3,200 +3,33 @@
 // Spawning systems for agent autonomy pipelines.
 //
 // Public: spawn_autonomy_pipeline, spawn_passthrough_pipeline, spawn_odom_frames
-// Private: build_separate_pipeline (shared helper), level_from_str
 
-use avian3d::prelude::Gravity;
 use bevy::prelude::*;
-use helios_core::mapping::NoneMapper;
 use helios_core::types::FrameHandle;
-use helios_runtime::config::MapperConfig;
+use helios_runtime::config::ControllerStateSourceConfig;
 use helios_runtime::estimation::GroundTruthPassthrough;
 use helios_runtime::pipeline::PipelineBuilder;
-use helios_runtime::stage::PipelineLevel;
 use helios_runtime::validation::validate_autonomy_config;
 use std::collections::HashMap;
 
 use crate::prelude::*;
-use crate::simulation::core::components::{AgentTopicNames, ControllerStateSource, SensorMailbox};
+use crate::simulation::core::components::{ControllerStateSource, SensorMailbox};
 use crate::simulation::plugins::autonomy::components::{
-    ControlPipelineComponent, EstimatorComponent, MapperComponent, ModuleTimer, OdomFrameOf,
+    ControlPipelineComponent, EstimatorComponent, MapperComponent, OdomFrameOf,
     PathFollowingComponent, PathFollowingOutputComponent,
 };
 use crate::simulation::registry::{
-    AutonomyRegistry, ControllerBuildContext, EstimatorBuildContext, MapperBuildContext,
-    PathFollowerBuildContext, PlannerBuildContext, SlamBuildContext,
+    AutonomyRegistry, EstimatorBuildContext, SlamBuildContext,
 };
-use helios_core::planning::types::PlannerGoal;
-use helios_runtime::config::ControllerStateSourceConfig;
 
-// =========================================================================
-// == Helpers ==
-// =========================================================================
-
-fn level_from_str(s: &str) -> PipelineLevel {
-    match s {
-        "global" => PipelineLevel::Global,
-        "local" => PipelineLevel::Local,
-        other => PipelineLevel::Custom(other.to_string()),
-    }
-}
-
-/// Shared helper for the `Separate` world-model path.
-///
-/// Builds mapper, controllers, planners, and goal from config; inserts all
-/// pipeline components onto the agent entity. The caller is responsible for
-/// constructing and passing the correct `EstimatorComponent` (real `EstimationCore`
-/// or `GroundTruthPassthrough`).
-fn build_separate_pipeline(
-    entity: Entity,
-    commands: &mut Commands,
-    map_cfg: Option<&MapperConfig>,
-    agent_config: &AgentConfig,
-    registry: &AutonomyRegistry,
-    est_component: EstimatorComponent,
-) {
-    let mut builder = PipelineBuilder::new();
-
-    if let Some(cfg) = map_cfg {
-        if let Some(rate) = cfg.get_timer_rate() {
-            commands.entity(entity).insert(ModuleTimer::from_hz(rate));
-        }
-        match registry.build_mapper(
-            cfg.get_kind_str(),
-            MapperBuildContext {
-                agent_entity: entity,
-                mapper_cfg: cfg.clone(),
-            },
-        ) {
-            Ok(m) => {
-                builder = builder.with_mapper(PipelineLevel::Local, m);
-            }
-            Err(e) => {
-                error!(
-                    "Mapper build failed for '{}': {}. Falling back to NoneMapper.",
-                    agent_config.name(),
-                    e
-                );
-                builder = builder.with_mapper(PipelineLevel::Local, Box::new(NoneMapper));
-            }
-        }
-    }
-
-    for (_key, ctrl_cfg) in &agent_config.autonomy_stack().controllers {
-        let kind = ctrl_cfg.get_kind_str();
-        let ctx = ControllerBuildContext {
-            agent_entity: entity,
-            controller_cfg: ctrl_cfg.clone(),
-            agent_config: agent_config.clone(),
-            dynamics_factories: registry.clone_dynamics_arc(),
-        };
-        match registry.build_controller(kind, ctx) {
-            Ok(ctrl) => {
-                builder = builder.with_controller(PipelineLevel::Local, ctrl);
-            }
-            Err(e) => {
-                error!(
-                    "Failed to build controller '{}' for agent '{}': {}",
-                    kind,
-                    agent_config.name(),
-                    e
-                );
-            }
-        }
-    }
-
-    for (_key, plan_cfg) in &agent_config.autonomy_stack().planners {
-        let kind = plan_cfg.get_kind_str();
-        let level = level_from_str(plan_cfg.get_level_str());
-        let ctx = PlannerBuildContext {
-            agent_entity: entity,
-            planner_cfg: plan_cfg.clone(),
-            level: level.clone(),
-        };
-        match registry.build_planner(kind, ctx) {
-            Ok(planner) => {
-                builder = builder.with_planner(level, planner);
-            }
-            Err(e) => {
-                error!(
-                    "Failed to build planner '{}' for agent '{}': {}",
-                    kind,
-                    agent_config.name(),
-                    e
-                );
-            }
-        }
-    }
-
-    if let Some(pf_cfg) = &agent_config.autonomy_stack().path_following {
-        let kind = pf_cfg.get_kind_str();
-        let ctx = PathFollowerBuildContext {
-            agent_entity: entity,
-            path_following_cfg: pf_cfg.clone(),
-        };
-        match registry.build_path_follower(kind, ctx) {
-            Ok(pf) => {
-                builder = builder.with_path_follower(pf);
-            }
-            Err(e) => {
-                error!(
-                    "Failed to build path follower '{}' for agent '{}': {}",
-                    kind,
-                    agent_config.name(),
-                    e
-                );
-            }
-        }
-    }
-
-    let goal_iso = agent_config.goal_pose.to_isometry();
-    builder = builder.with_goal(PlannerGoal::WorldPose(goal_iso));
-
-    let ctrl_state_source = agent_config
-        .autonomy_stack()
-        .controllers
-        .values()
-        .next()
-        .map(|c| c.state_source())
-        .unwrap_or_default();
-    let ctrl_state_source_component = match ctrl_state_source {
-        ControllerStateSourceConfig::GroundTruth => ControllerStateSource::GroundTruth,
-        ControllerStateSourceConfig::Estimated => ControllerStateSource::Estimated,
-    };
-
-    let agent_name = agent_config.name();
-    let (_, mapping, path_following, control) = builder.build().into_parts();
-    let mut entity_cmd = commands.entity(entity);
-    entity_cmd.insert((
-        est_component,
-        MapperComponent(Box::new(mapping)),
-        ControlPipelineComponent(control),
-        PathFollowingOutputComponent::default(),
-        ctrl_state_source_component,
-        SensorMailbox::default(),
-        AgentTopicNames {
-            active_waypoint: format!("/{}/planning/active_waypoint", agent_name),
-            odometry_estimated: format!("/{}/odometry/estimated", agent_name),
-            map_global: format!("/{}/map/global", agent_name),
-            map_local: format!("/{}/map", agent_name),
-        },
-    ));
-    if let Some(pf) = path_following {
-        entity_cmd.insert(PathFollowingComponent(pf));
-    }
-}
-
-// =========================================================================
-// == SPAWNING SYSTEMS ==
-// =========================================================================
+use super::spawn_helpers::build_separate_pipeline;
 
 /// Spawns the full autonomy pipeline for agents with real estimation (EKF/UKF).
-/// Queries `&Children` to build measurement models from sensor child entities.
 pub fn spawn_autonomy_pipeline(
     mut commands: Commands,
     agent_query: Query<(Entity, &SpawnAgentConfigRequest, &Children)>,
     measurement_model_query: Query<&MeasurementModel>,
-    gravity: Res<Gravity>,
+    gravity: Res<avian3d::prelude::Gravity>,
     registry: Res<AutonomyRegistry>,
 ) {
     let dynamics_factories = registry.clone_dynamics();
@@ -283,7 +116,6 @@ pub fn spawn_autonomy_pipeline(
                 };
                 match registry.build_slam(slam_cfg.get_kind_str(), ctx) {
                     Ok(system) => {
-                        let agent_name = agent_config.name();
                         let ctrl_state_source = agent_config
                             .autonomy_stack()
                             .controllers
@@ -311,15 +143,6 @@ pub fn spawn_autonomy_pipeline(
                             PathFollowingOutputComponent::default(),
                             ctrl_state_source_component,
                             SensorMailbox::default(),
-                            AgentTopicNames {
-                                active_waypoint: format!(
-                                    "/{}/planning/active_waypoint",
-                                    agent_name
-                                ),
-                                odometry_estimated: format!("/{}/odometry/estimated", agent_name),
-                                map_global: format!("/{}/map/global", agent_name),
-                                map_local: format!("/{}/map", agent_name),
-                            },
                         ));
                         if let Some(pf) = path_following {
                             entity_cmd.insert(PathFollowingComponent(pf));
@@ -345,8 +168,7 @@ pub fn spawn_autonomy_pipeline(
     }
 }
 
-/// Spawns the passthrough pipeline for mock-estimator profiles (MappingOnly, PlanningOnly, ControlOnly).
-/// Inserts `GroundTruthPassthrough` instead of a real EKF/UKF. Does not query `&Children`.
+/// Spawns the passthrough pipeline for mock-estimator profiles.
 pub fn spawn_passthrough_pipeline(
     mut commands: Commands,
     agent_query: Query<(Entity, &SpawnAgentConfigRequest)>,
@@ -387,11 +209,10 @@ pub fn spawn_passthrough_pipeline(
             }
 
             Some(WorldModelConfig::CombinedSlam { .. }) => {
-                let agent_name = agent_config.name();
                 warn!(
                     "Agent '{}' uses CombinedSlam but passthrough profile was requested. \
                      Inserting GT passthrough with empty mapper/control.",
-                    agent_name
+                    agent_config.name()
                 );
                 let ctrl_state_source = agent_config
                     .autonomy_stack()
@@ -414,12 +235,6 @@ pub fn spawn_passthrough_pipeline(
                     PathFollowingOutputComponent::default(),
                     ctrl_state_source_component,
                     SensorMailbox::default(),
-                    AgentTopicNames {
-                        active_waypoint: format!("/{}/planning/active_waypoint", agent_name),
-                        odometry_estimated: format!("/{}/odometry/estimated", agent_name),
-                        map_global: format!("/{}/map/global", agent_name),
-                        map_local: format!("/{}/map", agent_name),
-                    },
                 ));
                 if let Some(pf) = path_following {
                     entity_cmd.insert(PathFollowingComponent(pf));
