+++ helios_sim/src/simulation/plugins/debugging/ui/mod.rs
@@ -1,2 +1,3 @@
 pub mod legend;
 pub mod vehicle_hud;
+pub mod vehicle_hud_update;
